requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
catboost>=1.2.0
transformers>=4.30.0
torch>=2.0.0    
python-dateutil>=2.8.0
pydantic>=2.0.0